HTCONDOR-2721 Add Avg and Total TransferInput/Output MB attributes
  The Startd daemon ad will advertise 4 new attributes:
     AvgTransferInputMB, AvgTransferOutputMB,
     TotalTransferInputMB, TotalTransferOutputMB

  These attributes are updated when a starter sends its final update ad to the Startd.  Failed
  transfers do not typically report transfer bytes.

  This commit also changes the BytesSent and BytesRecvd job attributes so that they include
  bytes transferred via plugins.  As a consequence, the userlog report of bytes transferred
  will now also include bytes transferred via plugin.

  Several bugs with under and overcounting of plugin results were also fixed,
  These errors were mostly invisible to users since they were not included in the totals.

+++ src/condor_shadow.V6.1/baseshadow.h
@@ -236,7 +236,7 @@ class BaseShadow : public Service
 			zero bytes.
 			@return number of bytes sent over the network.
 		*/
-	virtual float bytesSent() { return 0.0; }
+	virtual uint64_t bytesSent() { return 0; }
 
 		/** The total number of bytes received over the network on
 			behalf of this job.
@@ -246,7 +246,7 @@ class BaseShadow : public Service
 			zero bytes.
 			@return number of bytes sent over the network.
 		*/
-	virtual float bytesReceived() { return 0.0; }
+	virtual uint64_t bytesReceived() { return 0; }
 
 	virtual void getFileTransferStats(ClassAd &upload_file_stats, ClassAd &download_file_stats) = 0;
 	ClassAd* updateFileTransferStats(const ClassAd& old_stats, const ClassAd &new_stats);
